+++ src/Web/Invest.MVC/Infrastructure/Core/Repositories/IForexRepository.cs
@@ -11,6 +11,6 @@ namespace Invest.MVC.Infrastructure.Core.Repositories
 
         void TakeSnapshot(Forex forex, DateTime date);
 
-        decimal GetExchangeRate(string fromCurrency, string toCurrency, DateTime date);
+        float GetExchangeRate(string fromCurrency, string toCurrency, DateTime date);
     }
 }
